Mod: project-wide review and refactor
Fix:
- cpu/worker: fix scheduler period unit confusion (was reusing
  kCpuMaxLoadPerCore as a millisecond constant)
- cpu/critical_loop: guard count<=0; use uint32_t to avoid signed
  overflow UB; cache PRNG dist
- cpu/stat & util/proc_stat: bound `%s` widths to prevent buffer
  overflow; RAII FILE* to fix leaks on error paths; parse
  /proc/<pid>/stat `comm` field by last ')' for names with spaces
- util/log: switch localtime to localtime_r for thread safety;
  LOG_FATAL now aborts instead of silently raising SIGTERM
- cli/options: failed validation now returns EXIT_FAILURE (was
  EXIT_SUCCESS); unknown options no longer silently skipped
- core/options: unify memory field as int64_t bytes (was int/MiB
  with inconsistent comments)
- util/rolling_sampler: O(1) incremental mean; fix
  "overwrite-twice" behavior when the ring first fills
- core/runtime: early-return when no module enabled; clarify
  NextSchedulingTime variable naming (ns, not us)
- cpu/manager_random_normal: error message now reports peak load
  instead of misleading avg target

Add:
- memory: spawn one-shot background thread for allocate+fill when
  block >= 32 MiB (matches README); catch std::bad_alloc
- main: use sigaction() with empty sa_mask on POSIX; signal
  handler is now async-signal-safe (atomic-only, no logging)

Mod:
- core/resource_manager: hoist last_scheduling_ into the base
  class to remove duplication in cpu/memory managers
- core/options: drop reverse dependency on cli/ via forward decl
- cli/version_string: drop singleton; lazy-init a static string
- CMake: target_* commands, find_package(Threads), Debug/Release
  separation, -Wall -Wextra -Wpedantic, optional static-link
  toggle (LOADMASTER_STATIC_LINK, Linux-only default)
- README: documented the background-allocator thread and build
  options; bumped minimum CMake to 3.12

Remove:
- util/waitgroup.h: unused, busy-loop Wait()
- util/win_util.cc: empty translation unit

Build verified: clean compile with -Werror -Wall -Wextra -Wpedantic
on GCC 13.3 (Release).

+++ src/cpu/manager_random_normal.h
@@ -15,7 +15,7 @@ class CpuResourceManagerRandomNormal final : public CpuResourceManager {
   virtual bool Init() override;
 
  private:
-  size_t point_idx_;
+  int point_idx_;
   std::vector<int> schedule_points_;
   util::NormalDistribution dist_;
   std::mt19937 generator_;
